Refactored signin component
