+++ frontend/src/app/components/AuthGuard.tsx
@@ -3,6 +3,7 @@
 import React, { useEffect } from "react";
 import { useAuth } from "../lib/AuthProvider";
 import { useRouter, usePathname } from "next/navigation";
+import CircularProgress from "@mui/material/CircularProgress";
 
 export default function AuthGuard({ children }: { children: React.ReactNode }) {
   const { user, loading } = useAuth();
@@ -18,8 +19,18 @@ export default function AuthGuard({ children }: { children: React.ReactNode }) {
   }, [user, loading, router, pathname]);
 
   if (loading && !user) {
-    // TODO replace with MUI loading component
-    return <div>Loading...</div>;
+    return (
+      <div>
+        <CircularProgress
+          size={40}
+          sx={{
+            position: "absolute",
+            color: "primary.light",
+          }}
+        />
+        Loading...
+      </div>
+    );
   }
 
   return <>{children}</>;
